# servers/core/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a core sever of PROMPT.",
        "title": "PROMPT Core API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/",
    "paths": {
...
        "/applications/{coursePhaseID}/assessment": {
            "put": {
                "description": "Batch update the status of multiple applications",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "applications"
                ],
                "summary": "Update applications status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course Phase UUID",
                        "name": "coursePhaseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status update",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/coursePhaseParticipationDTO.UpdateCoursePhaseParticipationStatus"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/applications/{coursePhaseID}/{courseParticipationID}/assessment": {
            "put": {
                "description": "Update the assessment for an application",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "applications"
                ],
                "summary": "Update application assessment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course Phase UUID",
                        "name": "coursePhaseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Course Participation UUID",
                        "name": "courseParticipationID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Assessment to update",
                        "name": "assessment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/applicationDTO.PutAssessment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "applicationDTO.PutAssessment": {
            "type": "object",
            "properties": {
                "passStatus": {
                    "$ref": "#/definitions/db.PassStatus"
                },
                "restrictedData": {
                    "$ref": "#/definitions/meta.MetaData"
                },
                "score": {
                    "type": "integer"
                }
            }
        },
...
        "coursePhaseParticipationDTO.UpdateCoursePhaseParticipationStatus": {
            "type": "object",
            "properties": {
                "courseParticipationIDs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "passStatus": {
                    "$ref": "#/definitions/db.PassStatus"
                }
            }
        },
...
        "db.PassStatus": {
            "type": "string",
            "enum": [
                "passed",
                "failed",
                "not_assessed"
            ],
            "x-enum-varnames": [
                "PassStatusPassed",
                "PassStatusFailed",
                "PassStatusNotAssessed"
            ]
        },
...
        "meta.MetaData": {
            "type": "object",
            "additionalProperties": true
        },
...
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Something went wrong"
                }
            }
        }
    },
...
}